+++ apps/web/components/ChatPanel/BoostChatPanel/index.ts
@@ -1,0 +1,4 @@
+import dynamic from 'next/dynamic'
+export const BoostChatPanel = dynamic(() => import('./BoostChatPanel'), {
+  ssr: false,
+})
