+++ apps/web/components/ChatPanel/BoostChatPanel/types.d.ts
@@ -1,4 +1,4 @@
-export type ChatPanelConfig = {
+export type BoostChatPanelConfig = {
   chatPanel?: {
     header?: {
       filters?: {
