+++ apps/web/components/ChatPanel/UkraineChatPanel/index.ts
@@ -1,0 +1,4 @@
+import dynamic from 'next/dynamic'
+export const UkraineChatPanel = dynamic(() => import('./UkraineChatPanel'), {
+  ssr: false,
+})
