+++ apps/web/components/ChatPanel/WatsonChatPanel/index.ts
@@ -1,0 +1,4 @@
+import dynamic from 'next/dynamic'
+export const WatsonChatPanel = dynamic(() => import('./WatsonChatPanel'), {
+  ssr: false,
+})
