+++ apps/web/components/real.ts
@@ -36,7 +36,6 @@ export * from './Section/Section'
 export * from './HeadWithSocialSharing/HeadWithSocialSharing'
 export * from './TellUsAStory/TellUsAStory'
 export * from './ContactUs/ContactUs'
-export * from './ChatPanel'
 export * from './AnchorNavigation/AnchorNavigation'
 export * from './Main/Main'
 export * from './Select/Select'
@@ -76,4 +75,4 @@ export * from './NewLinks/NewLinks'
 export * from './NewsItems/NewsItems'
 export * from './GridItems/GridItems'
 export * from './LifeEventsSection/LifeEventsSection'
-export * from './UkraineChatPanel/UkraineChatPanel'
+export * from './ChatPanel'
