+++ apps/web/screens/Project/components/ProjectChatPanel/config.ts
@@ -1,0 +1,7 @@
+export const config = {
+  '3NFDdRCIe5RVnlSEbHxoTT': {
+    integrationID: 'b1a80e76-da12-4333-8872-936b08246eaa',
+    region: 'eu-gb',
+    serviceInstanceID: 'bc3d8312-d862-4750-b8bf-529db282050a',
+  },
+}
