+++ apps/web/screens/Project/components/ProjectChatPanel/index.ts
@@ -1,0 +1,1 @@
+export * from './ProjectChatPanel'
