+++ apps/web/screens/Project/components/ProjectHeader/index.ts
@@ -1,0 +1,1 @@
+export * from './ProjectHeader'
