+++ apps/web/screens/Project/components/ProjectWrapper/index.ts
@@ -1,0 +1,1 @@
+export * from './ProjectWrapper'
